feat: add basic method for the bubble sorter
As the idea is to sort the elements in an array, a variable was used to
reference the value in an element. If the value of the current element is
larger than the next then the "swap" occurs. This swap uses the reference
value to change the value of the element for the specified index.