fix: end comparison once it reach the last value
The method successfully iterates through the element but continues to do
so even if it reach the last element of the array. To prevent this from
happening, a conditional statement was added. This was to check if there
is still an element to be compared by checking if the value returns nil.