fix: fix unexpected end of comparison
Testing the code, it was found out that only one pass happened throughout
the method. Meaning, the code ended unexpectedly because the iteration
reaches the last element of the array.

Checking the array, only the largest number was sorted to the very end.
To fix this, a loop was implemented so that each time an iteration is
completed a value will be incremented until that value is greater than
the array length.