fix: use a different loop to check if sorted
Using a counter that checks for each completed iteration/pass will either
make the sorter longer or shorter to complete.

Like a  nearly sorted array or an array that needs only one pass to
complete, it will mean that the method will continue the code execution
until the counter reaches the set value.

So a pass_count was implemented so that each pass will be counted every
time a swap is not executed. A pass_count reaching the "length of an
array minus 1" will mean the array was already sorted.